add weibo amount to myProfile

+++ src/main/java/com/up/weibo/service/IWeiboService.java
@@ -7,5 +7,5 @@ import com.up.model.WeiboEntity;
 public interface IWeiboService {
 	public void publish(String username, String content);
 	public List<WeiboEntity> getWeibo(String username);
-	public int getWeiboAmount(String username);
+	public Long getWeiboAmount(String username);
 }
